testing/web-platform/mozilla/tests/css/cssom/moz-transform-exists.html

<!doctype html>
<title>CSS Test: MozTransform exists in Gecko for compat reasons</title>
<meta charset="utf-8">
<link rel="author" title="Emilio Cobos Ãlvarez" href="mailto:emilio@crisal.io">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
test(function() {
  let el = document.documentElement;
  assert_true("MozTransform" in el.style, "MozTransform exists");
  el.style.MozTransform = "scale(1)";
  assert_not_equals(el.MozTransform, "", "MozTransform sets the value");
  assert_equals(el.transform, el.MozTransform, "MozTransform is an alias of transform");
});
</script>
